+++ src/colors.js
@@ -1,3 +1,5 @@
+export const black = '#000000';
+export const white = '#ffffff';
 export const backgroundColor = '#222222';
 export const titleColor = '#E3A44D';
 export const grandColor = '#6A6588';
